simple lighting

+++ src/Vector.kt
@@ -6,7 +6,7 @@ class Vector(var x:Double, var y:Double, var z:Double)
 fun vctrSum(a:Vector,b:Vector):Vector{
     return Vector(a.x+b.x,a.y+b.y,a.z+b.z)
 }
-fun vctrScale(a:Vector, b:Int):Vector{
+fun vctrScale(a:Vector, b:Double):Vector{
     return Vector(a.x*b,a.y*b,a.z*b);
 }
 fun vctrDot(a:Vector,b:Vector):Double{
